파워 수정

+++ Assets/Scripts/Player/P_Bullet.cs
@@ -8,7 +8,7 @@ public class P_Bullet : MonoBehaviour
     public float speed = 10f;
 
     // 파워
-    public float power = 1f;
+    public int power = 1;
 
     private void Start()
     {
